--- a/Assets/Prefabs/Sensors/InstanceSegmentationSensor.prefab.meta
+++ b/Assets/Prefabs/Sensors/InstanceSegmentationSensor.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2ae3265faf047844bb8f190a449c00c4
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: