--- a/Assets/Resources/RuntimeSettings.asset
+++ b/Assets/Resources/RuntimeSettings.asset
@@ -44,3 +44,4 @@ MonoBehaviour:
   - {fileID: 8926320593265141670, guid: 823225f8334c708988ff56b6fdf3c722, type: 3}
   - {fileID: 6798236283089428316, guid: b916b47d1234714418d5d1b2d71d2f8c, type: 3}
   - {fileID: 3075028223651123413, guid: c14ccc8ff9f29b3438cdd0f4bcebd180, type: 3}
+  - {fileID: 959914374408348488, guid: 2ae3265faf047844bb8f190a449c00c4, type: 3}