--- a/Assets/Scripts/Sensors/InstanceSegmentationSensor.cs.meta
+++ b/Assets/Scripts/Sensors/InstanceSegmentationSensor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cdd86435cb5d25e459a2ffe2da81ec82
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: