--- a/Assets/Shaders/InstanceSegmentationPass.hlsl.meta
+++ b/Assets/Shaders/InstanceSegmentationPass.hlsl.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 627ccdf40afaa714e8f88e204bfccb67
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: