--- a/Packages/com.unity.render-pipelines.high-definition/Editor/Material/Lit/ShaderGraph/HDLitMasterNode.cs
+++ b/Packages/com.unity.render-pipelines.high-definition/Editor/Material/Lit/ShaderGraph/HDLitMasterNode.cs
@@ -916,6 +916,13 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                 value = new Color(0.0f, 0.0f, 0.0f, 1.0f)
             });
 
+            collector.AddShaderProperty(new ColorShaderProperty()
+            {
+                overrideReferenceName = "_InstanceColor",
+                hidden = true,
+                value = new Color(0.0f, 0.0f, 0.0f, 1.0f)
+            });
+
             base.CollectShaderProperties(collector, generationMode);
         }
     }