… Assets/Scripts/Sensors/InstanceSegmentationSensor.cs