Based on Semantic Segmentation Sensor, created Instance Segmentation Sensor type. Assigns random color to object instances of specific categories (eg cars, pedestrians etc.). For users working with downloaded content, this will work only partially because in the downloaded content, the new shader pass isn't known, resulting in only objects rendering that have their materials outside the downloaded packages (this is true for AI vehicles and pedestrians). Once LG takes this change and rebuilds their content, this will be fixed.