@martins-mozeiko 2019-11-08
I'm not sure change on line 318 in SimulatorManager.cs from GetSharedMaterials to GetMaterials is a good idea.

The way how Unity renders is by batching - if multiple objects use same material instance they will be rendered with one drawcall. By making each material unique between similar objects your change prevents doing this. Unity needs to issue separate draw calls, because each object will have unique material.

Having drawcall per instance, not per material is a significant performance hit. On my machine for BorregasAve it drops performance from ~60 to 40 fps (with some sensors on the car). Without any sensors performance drops from 150 to 100 fps.

I'll think about how this can be changed. I understand that materials need to be different because each has different color assigned. Not sure how to change it. Technically Unity should be able to put this in instance specific constant buffer, but I don't know HDRP implementation details to explain how exactly this should be done. I think currently LitProperties.hlsl define properties only per material - `CBUFFER_START(UnityPerMaterial)` Per instance properties probably are somewhere else.