@KonstantinKohl 2019-11-11
@martins-mozeiko I wasn't aware of this implication, thanks for the heads-up.

Thinking of it, I will go on and change the code such that only for objects that ask for instance segmentation (AllowInstanceSegmentation), I will make the materials unique and keep the original code in place for the rest. In our case, only few objects will need instance segmentation (cars, pedestrians, traffic signs and traffic lights), while the majority of the world won't, so I trade a bit of performance against making the system work for us.

I'll submit my changes to this branch so your team can decide if this works for you.