@RamoramaInteractive 2020-03-08
I want to have Instance Segmentation Sensor type. I know, the performance of the simulator would be very low. What should I change to enable the Instance Segmentation Sensor type?