@rongguodong 2020-03-09
We will support instance segmentation soon.