@EricBoiseLGSVL 2020-04-10
@KonstantinKohl We had to make some major changes to make this work in simulator.  I am closing this but your work is merged in the feature.  Thank you for your hard work and this feature is now released.  The segmentation camera has parameters to enable this for each tag type.